+++ src/plugins/data_source_management/public/components/data_source_view/data_source_view.test.tsx
@@ -12,25 +12,39 @@ import {
   getSingleDataSourceResponse,
   mockErrorResponseForSavedObjectsCalls,
   mockResponseForSavedObjectsCalls,
+  mockManagementPlugin,
 } from '../../mocks';
 import { render } from '@testing-library/react';
+import * as utils from '../utils';
 
 describe('DataSourceView', () => {
   let component: ShallowWrapper<any, Readonly<{}>, React.Component<{}, {}, any>>;
   let client: SavedObjectsClientContract;
   const { toasts } = notificationServiceMock.createStartContract();
+  const mockedContext = mockManagementPlugin.createDataSourceManagementContext();
+  const uiSettings = mockedContext.uiSettings;
 
   beforeEach(() => {
+    jest.clearAllMocks();
     client = {
       get: jest.fn().mockResolvedValue([]),
     } as any;
     mockResponseForSavedObjectsCalls(client, 'get', getSingleDataSourceResponse);
   });
 
-  it('should render normally with local cluster not hidden', () => {
+  it('should render normally with local cluster not hidden', async () => {
+    spyOn(uiSettings, 'get').and.returnValue('test1');
     component = shallow(
-      <DataSourceView fullWidth={false} selectedOption={[{ id: 'test1', label: 'test1' }]} />
+      <DataSourceView
+        fullWidth={false}
+        selectedOption={[{ id: 'test1', label: 'test1' }]}
+        uiSettings={uiSettings}
+      />
     );
+    await component.instance().componentDidMount!();
+    expect(uiSettings.get).toBeCalledWith('defaultDataSource', null);
+    expect(component.state('selectedOption')).toEqual([{ id: 'test1', label: 'test1' }]);
+    expect(component.state('defaultDataSource')).toEqual('test1');
     expect(component).toMatchSnapshot();
     expect(toasts.addWarning).toBeCalledTimes(0);
   });
@@ -38,7 +52,7 @@ describe('DataSourceView', () => {
     const container = render(
       <DataSourceView fullWidth={false} selectedOption={[{ id: 'test1', label: 'test1' }]} />
     );
-    const button = await container.findByTestId('dataSourceViewContextMenuHeaderLink');
+    const button = await container.findByTestId('dataSourceViewrButton');
     button.click();
     expect(container).toMatchSnapshot();
   });
@@ -56,9 +70,8 @@ describe('DataSourceView', () => {
     expect(toasts.addWarning).toBeCalledTimes(0);
   });
   it('should call notification warning when there is data source fetch error', async () => {
-    jest.mock('../utils', () => ({
-      getDataSourceById: jest.fn(),
-    }));
+    spyOn(utils, 'getDataSourceById').and.returnValue(undefined);
+    spyOn(utils, 'handleDataSourceFetchError').and.returnValue('');
     component = shallow(
       <DataSourceView
         fullWidth={false}
@@ -67,9 +80,10 @@ describe('DataSourceView', () => {
         notifications={toasts}
       />
     );
+    await component.instance().componentDidMount!();
     expect(component).toMatchSnapshot();
     mockErrorResponseForSavedObjectsCalls(client, 'get');
-    expect(toasts.addWarning).toBeCalledTimes(1);
-    expect(toasts.addWarning).toBeCalledWith(`Data source with id test1 is not available`);
+    expect(utils.getDataSourceById).toHaveBeenCalled();
+    expect(utils.handleDataSourceFetchError).toHaveBeenCalled();
   });
 });