+++ src/plugins/data_source_management/public/components/data_source_view/data_source_view.tsx
@@ -5,22 +5,41 @@
 
 import React from 'react';
 import { i18n } from '@osd/i18n';
-import { EuiPopover, EuiButtonEmpty, EuiButtonIcon, EuiContextMenu } from '@elastic/eui';
+import {
+  EuiButtonEmpty,
+  EuiPopover,
+  EuiContextMenuPanel,
+  EuiPanel,
+  EuiSelectable,
+  EuiSwitch,
+} from '@elastic/eui';
 import { SavedObjectsClientContract, ToastsStart } from 'opensearch-dashboards/public';
+import { IUiSettingsClient } from 'src/core/public';
 import { DataSourceOption } from '../data_source_menu/types';
-import { getDataSourceById } from '../utils';
-import { MenuPanelItem } from '../../types';
+import { getDataSourceById, getDataSources, handleDataSourceFetchError } from '../utils';
+import './_data_source_view.scss';
+import { DataSourceOptionItem } from '../data_source_option';
 
 interface DataSourceViewProps {
   fullWidth: boolean;
   selectedOption: DataSourceOption[];
   savedObjectsClient?: SavedObjectsClientContract;
   notifications?: ToastsStart;
+  uiSettings?: IUiSettingsClient;
+}
+
+interface DataSourceOptionDisplay extends DataSourceOption {
+  disabled?: boolean;
+  checked?: string;
 }
 
 interface DataSourceViewState {
-  selectedOption: DataSourceOption[];
+  selectedOption: DataSourceOptionDisplay[];
   isPopoverOpen: boolean;
+  defaultDataSource: string | null;
+  allDataSources: any;
+  checked: boolean;
+  dataSourceOptionDisplay: DataSourceOptionDisplay[];
 }
 
 export class DataSourceView extends React.Component<DataSourceViewProps, DataSourceViewState> {
@@ -32,6 +51,10 @@ export class DataSourceView extends React.Component<DataSourceViewProps, DataSou
     this.state = {
       isPopoverOpen: false,
       selectedOption: this.props.selectedOption ? this.props.selectedOption : [],
+      defaultDataSource: null,
+      allDataSources: [],
+      checked: true,
+      dataSourceOptionDisplay: [],
     };
   }
 
@@ -45,29 +68,38 @@ export class DataSourceView extends React.Component<DataSourceViewProps, DataSou
 
     const option = selectedOption[0];
     const optionId = option.id;
+    const defaultDataSource = this.props.uiSettings?.get('defaultDataSource', null) ?? null;
+    const allDataSources = await getDataSources(this.props.savedObjectsClient!);
     if (!option.label) {
       try {
-        const title = (await getDataSourceById(optionId, this.props.savedObjectsClient)).title;
+        const title = (await getDataSourceById(optionId, this.props.savedObjectsClient!)).title;
         if (!title) {
-          this.props.notifications.addWarning(
-            i18n.translate('dataSource.fetchDataSourceError', {
-              defaultMessage: `Data source with id ${optionId} is not available`,
-            })
-          );
+          handleDataSourceFetchError(optionId, this.props.notifications!);
         } else {
           if (!this._isMounted) return;
           this.setState({
-            selectedOption: [{ id: optionId, label: title }],
+            selectedOption: [{ id: optionId, label: title, checked: 'on', disabled: true }],
+            dataSourceOptionDisplay: [
+              { id: optionId, label: title, checked: 'on', disabled: true },
+            ],
+            defaultDataSource,
+            allDataSources,
           });
+          return;
         }
       } catch (error) {
-        this.props.notifications.addWarning(
-          i18n.translate('dataSource.fetchDataSourceError', {
-            defaultMessage: `Failed to fetch data source due to ${error}`,
-          })
-        );
+        handleDataSourceFetchError(optionId, this.props.notifications!, error);
       }
     }
+    this.setState({
+      ...this.state,
+      selectedOption: [{ id: optionId, label: option.label, checked: 'on', disabled: true }],
+      dataSourceOptionDisplay: [
+        { id: optionId, label: option.label, checked: 'on', disabled: true },
+      ],
+      defaultDataSource,
+      allDataSources,
+    });
   }
 
   onClick() {
@@ -78,67 +110,79 @@ export class DataSourceView extends React.Component<DataSourceViewProps, DataSou
     this.setState({ ...this.state, isPopoverOpen: false });
   }
 
-  getPanels() {
-    let items: MenuPanelItem[] = [];
-    if (this.state.selectedOption) {
-      items = this.state.selectedOption.map((option) => {
+  onChange(e) {
+    const checked = e.target.checked;
+    let dataSourcesAggreated = this.state.allDataSources.map((dataSource) => {
+      if (dataSource.id === this.state.selectedOption[0].id) {
         return {
-          name: option.label,
+          id: this.state.selectedOption[0].id,
+          label: this.state.selectedOption[0].label,
+          checked: 'on',
           disabled: true,
         };
-      });
-    }
-
-    const panels = [
-      {
-        id: 0,
-        title: 'Selected data source',
-        items,
-      },
-    ];
-
-    return { panels };
+      } else {
+        return { id: dataSource.id, label: dataSource.title!, disabled: true };
+      }
+    });
+    dataSourcesAggreated = checked
+      ? dataSourcesAggreated.filter((item) => item.checked === 'on')
+      : dataSourcesAggreated;
+    this.setState({ checked, dataSourceOptionDisplay: dataSourcesAggreated });
   }
 
   render() {
-    const { panels } = this.getPanels();
-
+    const options = this.state.dataSourceOptionDisplay;
+    const label = this.state.selectedOption[0].label;
     const button = (
-      <EuiButtonIcon
-        iconType="iInCircle"
-        display="empty"
-        aria-label="Next"
-        onClick={this.onClick.bind(this)}
-      />
-    );
-    return (
       <>
         <EuiButtonEmpty
-          className="euiHeaderLink"
-          data-test-subj="dataSourceViewContextMenuHeaderLink"
-          aria-label={i18n.translate('dataSourceView.dataSourceOptionsButtonAriaLabel', {
-            defaultMessage: 'dataSourceViewMenuButton',
+          className="dataSourceViewButton"
+          data-test-subj="dataSourceViewrButton"
+          onClick={this.onClick.bind(this)}
+          aria-label={i18n.translate('dataSourceView.dataSourceOptionsViewAriaLabel', {
+            defaultMessage: 'dataSourceViewButton',
           })}
           iconType="database"
           iconSide="left"
           size="s"
-          disabled={true}
         >
-          {this.state.selectedOption && this.state.selectedOption.length > 0
-            ? this.state.selectedOption[0].label
-            : ''}
+          {label}
         </EuiButtonEmpty>
-        <EuiPopover
-          id={'dataSourceViewContextMenuPopover'}
-          button={button}
-          isOpen={this.state.isPopoverOpen}
-          closePopover={this.closePopover.bind(this)}
-          panelPaddingSize="none"
-          anchorPosition="downLeft"
-        >
-          <EuiContextMenu initialPanelId={0} panels={panels} />
-        </EuiPopover>
       </>
+    );
+
+    return (
+      <EuiPopover
+        id={'dataSourceViewPopover'}
+        button={button}
+        isOpen={this.state.isPopoverOpen}
+        closePopover={this.closePopover.bind(this)}
+        panelPaddingSize="none"
+        anchorPosition="downLeft"
+      >
+        <EuiContextMenuPanel title={`DATA SOURCE 1/${this.state.allDataSources.length}`}>
+          <EuiPanel color="transparent" paddingSize="s" className="panelWidth">
+            <EuiSelectable
+              options={options}
+              renderOption={(option) => (
+                <DataSourceOptionItem
+                  item={option}
+                  defaultDataSource={this.state.defaultDataSource}
+                />
+              )}
+            >
+              {(list) => list}
+            </EuiSelectable>
+            <div className="euiContextMenuItem">
+              <EuiSwitch
+                label={`Used on this page (1)`}
+                checked={this.state.checked}
+                onChange={(e) => this.onChange(e)}
+              />
+            </div>
+          </EuiPanel>
+        </EuiContextMenuPanel>
+      </EuiPopover>
     );
   }
 }